Add files via upload
The Leave Management System is an Lan based or online application which can be used throughout the corporate organization . This project system can be used to automate the workflow of leave applications and their approvals in corporate world. This system have features like notification of email, leave cancellation, automated leave approval and report generation based on the requested data by the user.